moves all everies and modifiers to their own files
