+++ lib/src/everies/every_date_validator_list_mixin.dart
@@ -1,0 +1,9 @@
+import '../date_validator.dart';
+import 'every_date_validator.dart';
+
+/// Mixin that represents a list of [EveryDateValidator].
+mixin EveryDateValidatorListMixin<E extends EveryDateValidator>
+    on DateValidatorListMixin<E> {
+  /// List for all of the [everies] that will be used to generate the date.
+  List<EveryDateValidator> get everies => [...this];
+}
