+++ lib/src/extensions.dart
@@ -1,6 +1,9 @@
 import 'package:time/time.dart';
 
-import '../due_date.dart';
+import 'date_validator.dart';
+import 'due_date.dart';
+import 'enums.dart';
+import 'everies/everies.dart';
 
 /// Extension methods to get the day of the year of a [DateTime].
 extension DayInYear on DateTime {
